design modificat

+++ Proiect ASP final/Controllers/AdresaController.cs
@@ -1,26 +1,46 @@
-﻿using Proiect_ASP_final.Models;
+﻿using Microsoft.AspNet.Identity;
+using Proiect_ASP_final.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Web;
 using System.Web.Mvc;
 
 namespace Proiect_ASP_final.Controllers
-{   
+{
     [RequireHttps]
     public class AdresaController : Controller
     {
-        // IDEE PENTRU ADRESA
-        // Gasisem un API cu toate tarile -> zonele -> orasele din ele pe care-l putem folosi (sa nu introduca utilizatorul aiurea)
         private Models.ApplicationDbContext db = new ApplicationDbContext();
 
-        // GET: Adresa
+        [Authorize(Roles = "User,Admin")]
         public ActionResult Index()
         {
-            var adrese = from a in db.Adrese
-                         select a;
+            var userId = User.Identity.GetUserId();
 
-            ViewBag.adrese = adrese;
+            ViewBag.isAdmin = false;
+            if (User.IsInRole("Admin"))
+                ViewBag.isAdmin = true;
+
+            if (ViewBag.isAdmin)
+            {
+                var adrese = from a in db.Adrese
+                             select a;
+                ViewBag.adrese = adrese;
+                if (adrese.Any())
+                    ViewBag.existaAdrese = true;
+            }
+            else
+            {
+                var adrese = from a in db.Adrese
+                             where a.idUtilizator == userId
+                             select a;
+                ViewBag.adrese = adrese;
+                if (adrese.Any())
+                    ViewBag.existaAdrese = true;
+            }
+
 
             if (TempData["mesaj"] != null)
                 ViewBag.mesaj = TempData["mesaj"];
@@ -28,39 +48,52 @@ namespace Proiect_ASP_final.Controllers
             return View();
         }
 
-        /*
-		Aici ar trebui sa avem un index separat cu id ca parametru sa dam doar 
-        adresele unui utilizator (nu ne trebuie toate decat cand avem de-a face cu un admin)
-		dar nu avem nevoie atata timp cat nu avem utilizatori
-        public ActionResult Index(int i)
-        {
-            var adrese = from a in db.Adrese
-                         where a.idUtilizator == i
-                         select a;
-
-            ViewBag.adrese = adrese;
-
-            if (TempData["mesaj"] != null)
-                ViewBag.mesaj = TempData["mesaj"];
-
-            return View();
-        }
-        */
-
+        [Authorize(Roles = "User,Admin")]
         public ActionResult Afisare(int id)
         {
+            var isAdmin = false;
+            if (User.IsInRole("Admin"))
+                isAdmin = true;
+
             Adresa adresaCautata = db.Adrese.Find(id);
+
+            if (adresaCautata.idUtilizator != User.Identity.GetUserId() && isAdmin == false)
+            {
+                TempData["mesaj"] = "Nu aveți permisiunea să vizualizați această adresă";
+                return RedirectToAction("Index");
+            }
+
+            var comenzi = (from c in db.Comenzi
+                           where c.idAdresa == id
+                           select c).OrderBy(comanda => comanda.dataPlasare);
+
+            ViewBag.comenzi = comenzi;
+            if (comenzi.Any())
+                ViewBag.areComenzi = true;
+
             return View(adresaCautata);
         }
 
+        [Authorize(Roles = "Admin,User")]
         public ActionResult Editare(int id)
         {
+            var isAdmin = false;
+            if (User.IsInRole("Admin"))
+                isAdmin = true;
+
             Adresa adresaDeEditat = db.Adrese.Find(id);
+
+            if (adresaDeEditat.idUtilizator != User.Identity.GetUserId() && isAdmin == false)
+            {
+                TempData["mesaj"] = "Nu aveți permisiunea să vizualizați această adresă";
+                return RedirectToAction("Index");
+            }
 
             return View("FormEditare", adresaDeEditat);
         }
 
         [HttpPut]
+        [Authorize(Roles = "Admin,User")]
         public ActionResult Editare(int id, Adresa adresaActualizata)
         {
             Adresa adresaDeEditat = db.Adrese.Find(id);
@@ -92,6 +125,7 @@ namespace Proiect_ASP_final.Controllers
             }
         }
 
+        [Authorize(Roles = "User")]
         public ActionResult Adaugare()
         {
             Adresa adresa = new Adresa();
@@ -101,6 +135,8 @@ namespace Proiect_ASP_final.Controllers
         [HttpPost]
         public ActionResult Adaugare(Adresa adresaDeAdaugat)
         {
+            var userId = User.Identity.GetUserId();
+            adresaDeAdaugat.idUtilizator = userId;
 
             try
             {
@@ -123,15 +159,29 @@ namespace Proiect_ASP_final.Controllers
         }
 
         [HttpDelete]
+        [Authorize(Roles = "User,Admin")]
         public ActionResult Stergere(int id)
         {
+            var isAdmin = false;
+            if (User.IsInRole("Admin"))
+                isAdmin = true;
+
             try
             {
+                string userId = User.Identity.GetUserId();
                 Adresa adresaDeSters = db.Adrese.Find(id);
+                var comenzi = (from c in db.Comenzi
+                              where c.idUtilizator == userId
+                              select c).ToList<Comanda>();
+
+                if (adresaDeSters.idUtilizator != User.Identity.GetUserId() && isAdmin == false || comenzi.Any())
+                {
+                    TempData["mesaj"] = "Nu aveți permisiunea să ștergeți acestă adresă";
+                    return RedirectToAction("Index");
+                }
+
                 db.Adrese.Remove(adresaDeSters);
-
                 db.SaveChanges();
-
                 TempData["mesaj"] = "Adresa a fost stearsa cu succes!";
 
                 return RedirectToAction("Index");
